[Author: mrossetti]
This is it!  The final, as in 'last' fix for http://code.google.com/p/qsb-mac/issues/detail?id=18.  Dasher/hosted/Google Apps accounts are now supported.  The previous implementation suffered from a rare meltdown in the case where the user had the same password for their hosted as for their non-hosted account, didn't remember their password, and tried to set up enough times that a captcha was required FOR BOTH ACCOUNT possibilities.  What we do now is detect the possibility of a Google Apps account by examining the domain (right now just for it not being '@gmail.com') and present a checkbox  letting the user specify if they are setting up a Google Apps account or a regular Google account.  Fixed GoogleCorporaSource so that it pulls the dasher accounts from the GoogleApps account type.  Corrected a misspelling for how desired accounts are specified. Refactored accounts a bit to clean up the initialization process.

PRESUBMIT=passed
R=dmaclach
DELTA=1532  (328 added, 1034 deleted, 170 changed)
OCL=10508938
SCL=10538107


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@106 c7262d1c-4ebb-4cc0-bccf-18c293255191