[Author: mrossetti]
Picasaweb does not _require_ credentials in order to provide albums so I've changed the PicasawebSource so that if no password is provided it does a non-credentialled fetch.  It's not currently possible to set up a Google account without a password (it won't authenticate) so this is for future enhancement.  (Some design discussion needs to be had in order to allow such access.)

PRESUBMIT=passed
R=dmaclach,hawk
DELTA=6  (3 added, 0 deleted, 3 changed)
OCL=11030359
SCL=11030674


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@254 c7262d1c-4ebb-4cc0-bccf-18c293255191