[Author: mrossetti]
The problem shown in auto-completion is that the selection range was being set to the middle of a glyph.  Interestingly, the text view character counting considers an accented character as a single character but selection is thinking in terms of glyphs.  When the typed string was 'espé' the string length is given as 4.  Doing a -[NSString rangeOfString:options:] (See -[QSBTextField insertCompletion:forPartialWordRange:movement:isFinal:]) with the 'espé' string against a completion string of 'espécie.txt' returned a range of length 5.  This was the clue.  At the end of the insertCompletion function we restore the selection to 0,4 but that ends up in the middle of the é glyph.  I added code in the -[QSBTextField setSelectedRanges:affinity:stillSelecting: override] to adjust each range so that selections would not fall in the middle of glyphs: if it's an empty selection then the selection is placed just after the glyph, otherwise the start of the selection is adjusted to just in front of the first glyph and the end of the selection is adjusted to be just after the last glyph. 
Fixes http://code.google.com/p/qsb-mac/issues/detail?id=533.

PRESUBMIT=passed
R=dmaclach,hawk
DELTA=72  (53 added, 4 deleted, 15 changed)
OCL=11627063
SCL=11641970


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@405 c7262d1c-4ebb-4cc0-bccf-18c293255191