[Author: mrossetti]
Fix the search box so that it doesn't appear when there is a modal dialog being presented, such as for the Google Docs Save As... dialog.  Also, force the Save As... dialog to the front by pulling the QSB application to the front.

PRESUBMIT=passed
R=dmaclach,hawk
DELTA=63  (13 added, 5 deleted, 45 changed)
OCL=11722015
SCL=11725267


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@436 c7262d1c-4ebb-4cc0-bccf-18c293255191