[Author: mrossetti]
Implemented custom results view presentation.  Refactored the results view controllers a bit so that they all have a reference to the detail view so that the detail view can be suppressed whenever there is a custom results view.  The custom view is optional and, if used, is specified in the extension's plist by providing a dictionary with result type(s) as key(s) and a sub-dictionary as the value.  The sub-dictionary contains two items specifying the name of the nib and the name of the nib's view controller class with which the result will be presented.  (The keys for these dictionaries are given in QSBResultRowViewController .m.)  The view controller class should have a -[setResult:(HGSResult *)result] function, providing the main connection between the result and how the view wishes to present it.

PRESUBMIT=passed
R=hawk,dmaclach
DELTA=459  (344 added, 80 deleted, 35 changed)
OCL=11736998
SCL=11750888


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@441 c7262d1c-4ebb-4cc0-bccf-18c293255191