[Author: mrossetti]
Added a custom view for the stock quoter.  Note: Once the user shortcuts by selecting the quote to bring up the Google Finance web page for the selected stock the stock quote display will revert to the non-custom presentation because the result values needed for the custom presentation are not being cached.  (If I had it to do over again, I would probably wrote the source in Obj-C.)

PRESUBMIT=passed
R=hawk,dmaclach
DELTA=1875  (1773 added, 36 deleted, 66 changed)
OCL=11807956
SCL=11822062


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@451 c7262d1c-4ebb-4cc0-bccf-18c293255191