[Author: hawk]
A clipboard history search source. The source keeps a queue of the last 25 items on the clipboard. For string-based clipboard entries, you can search based on the contents of the string, or you can search for "clipboard" and pivot in to the result to get a list of the entire history. Strings (plain and RTF), URLs, and images (with nifty icon preview) are supported.

OCL=11935812
DELTA=638 (636 added, 0 deleted, 2 changed)
R=mrossetti,dmaclach
SCL=12008344


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@478 c7262d1c-4ebb-4cc0-bccf-18c293255191