[Author: hawk]
Allow any file:// URL that points to one of out known screensavers to be used to set a screensaver, such as one returned by the Spotlight source. Do so by removing use of proprietary result keys and instead checking if the URL points to one of the screensavers.

OCL=12221451
DELTA=77 (39 added, 11 deleted, 27 changed)
R=mrossetti
SCL=12233269


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@510 c7262d1c-4ebb-4cc0-bccf-18c293255191