[Author: mrossetti]
Add a new scoring approach.  Note that the old scoring approach is still in the product but will be removed before shipping Cobalt.  The old scoring approach can be activated by adding an array to QSB preferences with a key of 'HGSTermScoringSettingsKey' and a single boolean value at index 0 of 'YES'.  The 10 factors used in the new scoring algorithm can also be set using this array.  To see what these values are please visit the code for -[HGSMemorySearchSource performSearchOperation:].  If you want to experiment with the scoring algorithm please see the TermScoring project in mrossetti's experimental directory.

PRESUBMIT=passed
R=dmaclach
DELTA=1054  (1018 added, 2 deleted, 34 changed)
OCL=12328860
SCL=12370424


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@527 c7262d1c-4ebb-4cc0-bccf-18c293255191