[Author: mrossetti]
When a HOSTED account fails to authenticate it might be because it is a GA+ account so fall back to HOSTED_OR_GOOGLE and try again.  If that works, mark the account as needing to be forced in the future.  This is a temporary workaround for http://buganizer/issue?id=2076723 and will need to be reworked when the blocking bug, 1960732, is fixed.  Yes, it is possible to make this robust enough to survive the future fix of 1960732 but the effort doesn't seem worth the complex solution for the limited audience and such a short time period.

PRESUBMIT=passed
R=dmaclach
DELTA=108  (102 added, 1 deleted, 5 changed)
OCL=12990192
SCL=13018976


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@628 c7262d1c-4ebb-4cc0-bccf-18c293255191