[Author: dmaclach]
A small renaming from "delegates" to "controllers" as I found the inheritence
chain to be very confusing (a "delegate" inherited from a "controller"?),
and we had places calling setController that were taking "delegates".
This will break the build (Due to an intermixed change). I will fix it with
a followup CL (not that it matters right now with Speech broken).

PRESUBMIT=passed
R=mrossetti
DELTA=2416  (1172 added, 1181 deleted, 63 changed)
OCL=13755592
SCL=13756057


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@646 c7262d1c-4ebb-4cc0-bccf-18c293255191