[Author: dmaclach]
Refactored HGSResult into HGSResult/HGSScoredResult/HGSUnscoredResult.
Changed queries so that we are passing around HGSTokenizedStrings instead of
NSString which allows us to keep match index data.
Added support to python plugins for displaying notifications.
Changed tokenization so that it breaks words like firefox and photoshop instead
of depending on the scorer to take care of it.
Changed several APIs to use the "score" instead of "rank".
Changed scoring to move it back close to the old algorithm based on the fact
that we no longer need to do inner word matches (i.e. pshop is photoshop). The
removal of inner word matches greatly simplified life.
Removed HGSStringUtil.
Reworked HGSMixer to mix more efficiently, and stop appropriately.
Added some fields to HGSResults in the DebugWindow.
Added support for blacklisting paths to the Spotlight source to get rid of a
pile of spurious results.
Several other small tweaks.

This still requires some cleanup, but I needed to get it in.
Let me know if you want to discuss the scorer changes.
Apologies for the massive CL, but if you could review this ASAP I'd really
appreciate it.

Bugs fixed:
http://code.google.com/p/qsb-mac/issues/detail?id=832 : Allow access to Growl
from Python plugin
http://code.google.com/p/qsb-mac/issues/detail?id=883 : QSB enters double when
typing
http://code.google.com/p/qsb-mac/issues/detail?id=882 : When traversing folders,
subfolders are shown last
http://code.google.com/p/qsb-mac/issues/detail?id=865 : camelcaps interfering
with ranking
http://code.google.com/p/qsb-mac/issues/detail?id=860 : QSB fails to search for
any internet result
http://code.google.com/p/qsb-mac/issues/detail?id=790 : Result pane gets
distorted if we do an image search.
http://code.google.com/p/qsb-mac/issues/detail?id=738 : Scorer Should Not Recognize Overlapping Search Terms

Amongst others....

PRESUBMIT=passed
R=mrossetti
OCL=14293520
DELTA=4
SCL=14305182


git-svn-id: file:///Users/gqj/repos/qsb-mac-archive/p/trunk@723 c7262d1c-4ebb-4cc0-bccf-18c293255191